+++ src/main.c
@@ -46,7 +46,7 @@ free_text_info(void *data) {
 static void
 main_setup_tree_columns(GtkWidget *tree, enum CecupColumn col_act, enum CecupColumn col_path) {
     GActionMap *action_map = G_ACTION_MAP(cecup.application);
-    int32 side = GPOINTER_TO_INT(g_object_get_data(G_OBJECT(tree), "side"));
+    int8 side = (int8)GPOINTER_TO_INT(g_object_get_data(G_OBJECT(tree), "side"));
 
     if (g_action_map_lookup_action(action_map, "tree_dispatch") == NULL) {
         GSimpleAction *dispatch;
